# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: Browse available tasks
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, paused, completed, all]
            default: open
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Task marketplace page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task as an advertiser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/{taskId}/submissions:
    post:
      operationId: submitTask
      tags: [tasks]
      summary: Submit proof for a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /submissions:
    get:
      operationId: listSubmissions
      tags: [tasks]
      summary: List the signed-in user's submissions
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [mine, owned]
            default: mine
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, all]
            default: all
      responses:
        "200":
          description: Submission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /submissions/{submissionId}/review:
    post:
      operationId: reviewSubmission
      tags: [tasks]
      summary: Approve or reject a worker submission
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionReviewInput"
      responses:
        "200":
          description: Reviewed submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
    SubmissionId:
      name: submissionId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Task:
      type: object
      required: [id, title, advertiserId, advertiserName, category, description, instructions, proofType, requiresKyc, reward, totalBudget, remainingBudget, maxCompletions, completedCount, status, deadline, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        advertiserId: { type: string }
        advertiserName: { type: string }
        category: { type: string }
        description: { type: string }
        instructions: { type: string }
        proofType: { type: string, enum: [text, url, image, file] }
        requiresKyc: { type: boolean }
        reward: { type: number }
        totalBudget: { type: number }
        remainingBudget: { type: number }
        maxCompletions: { type: integer }
        completedCount: { type: integer }
        status: { type: string, enum: [open, paused, completed] }
        deadline: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    TaskInput:
      type: object
      required: [title, category, description, instructions, proofType, requiresKyc, maxCompletions, deadline]
      properties:
        title: { type: string, minLength: 3, maxLength: 120 }
        category: { type: string, minLength: 2, maxLength: 40 }
        description: { type: string, minLength: 10, maxLength: 1000 }
        instructions: { type: string, minLength: 10, maxLength: 2000 }
        proofType: { type: string, enum: [text, url, image, file] }
        requiresKyc: { type: boolean }
        maxCompletions: { type: integer, minimum: 1, maximum: 100000 }
        deadline: { type: string, format: date-time }
    TaskList:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    Submission:
      type: object
      required: [id, taskId, taskTitle, workerId, workerName, proof, status, reward, submittedAt]
      properties:
        id: { type: string }
        taskId: { type: string }
        taskTitle: { type: string }
        workerId: { type: string }
        workerName: { type: string }
        proof: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        reward: { type: number }
        reviewerNote: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
    SubmissionInput:
      type: object
      required: [proof]
      properties:
        proof: { type: string, minLength: 1, maxLength: 5000 }
    SubmissionReviewInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        reviewerNote: { type: string, maxLength: 1000 }
...