# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskProofType } from './taskProofType';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: string;
  title: string;
  advertiserId: string;
  advertiserName: string;
  category: string;
  description: string;
  instructions: string;
  proofType: TaskProofType;
  requiresKyc: boolean;
  reward: number;
  totalBudget: number;
  remainingBudget: number;
  maxCompletions: number;
  completedCount: number;
  status: TaskStatus;
  deadline: Date;
  createdAt: Date;
}