# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputProofType } from './taskInputProofType';

export interface TaskInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 2
     * @maxLength 40
     */
  category: string;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  description: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  instructions: string;
  proofType: TaskInputProofType;
  requiresKyc: boolean;
  /**
     * @minimum 1
     * @maximum 100000
     */
  maxCompletions: number;
  deadline: Date;
}